third_party/WebKit/LayoutTests/editing/selection/mouse/drag-selects-culled-inlines.html

<!DOCTYPE html>
<style>
.style1 { font-size: 14px; color: #CC3300; }
.text1 { font-size: 11px; color: #7f0000; }
</style>
<div id="test" class="style1">
<font>
<span class="text1">Some Text1 <span id="noRange" style="font-weight: 400"> <b>Some Text2</b></span><br>Some Text3<br><b id="lastLine">Some Text4</b><br></span>
</font>
</div>
<p id="description">
<div id="console">
</div>
</p>
<script src="../../../resources/js-test.js"></script>
<script>
    jsTestIsAsync = true;
    var selectionComplete = false;
    description('Test that flickering is observed on text selection when culled inline is selected');

    document.onselectionchange = function () {
        shouldBeFalse("window.getSelection().containsNode(document.getElementById('noRange'))");
        if (selectionComplete)
            finishJSTest();
    }

    if (window.testRunner && window.eventSender) {
        var lastLine = document.getElementById('lastLine');

        eventSender.mouseMoveTo(lastLine.offsetLeft + 10, lastLine.offsetTop + lastLine.offsetHeight);
        eventSender.mouseDown();

        // move to the position of culled inline.
        eventSender.mouseMoveTo(lastLine.offsetLeft + 10, lastLine.offsetTop - 1);
        eventSender.mouseUp();
        selectionComplete = true;
    }
</script>
