third_party/WebKit/LayoutTests/editing/selection/mouse/drag-selects-culled-inlines-expected.txt

Some Text1 Some Text2
Some Text3
Some Text4


Test that flickering is observed on text selection when culled inline is selected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS window.getSelection().containsNode(document.getElementById('noRange')) is false
PASS successfullyParsed is true

TEST COMPLETE
